Fix builds for projects that do not build .war archives

+++ build/build-war.sh
@@ -9,10 +9,13 @@ find ${DOCKER_BUILD_DIR}
 
 cp ci-tools/build/Dockerfile ${DOCKER_BUILD_DIR}/Dockerfile
 sed -i -e "s|BASEIMAGE|${ECR_REPO}/${BASE_IMAGE}|g" ${DOCKER_BUILD_DIR}/Dockerfile
-for artifact in $DOCKER_BUILD_DIR/artifact/*.war; do 
-  unzip -d "$DOCKER_BUILD_DIR/artifact/$(basename $artifact .war)" $artifact
-  rm -f $artifact
-done
+
+if [ "$(ls -A $DOCKER_BUILD_DIR/artifact/*.war)" ]; then 
+  for artifact in $DOCKER_BUILD_DIR/artifact/*.war; do 
+    unzip -d "$DOCKER_BUILD_DIR/artifact/$(basename $artifact .war)" $artifact
+    rm -f $artifact
+  done
+fi
 docker build --build-arg name=${ARTIFACT_NAME} --build-arg artifact_destination=${ARTIFACT_DEST_PATH} -t ${DOCKER_TARGET} ${DOCKER_BUILD_DIR}
 
 docker images
